add support for mixed class arrays, move code to a separate file
