+++ ui/lib/src/view/index.ts
@@ -3,4 +3,5 @@ export * from './controls';
 export * from './dialog';
 export * from './dialogs';
 export * from './snabbdom';
+export * from './snabbdomTemplate';
 export * from './miniBoard';
